Extract wiki interface to submodule
